fix directory structure
